+++ Source/Object/Mesh.h
@@ -7,25 +7,25 @@ class Mesh
 	friend Shape;
 
 private:
-	unsigned short int _countVertex = 0;
+	unsigned int _countVertex = 0;
 	float* _aVertex = nullptr;
 	float* _aNormal = nullptr;
 	float* _aTexCoord = nullptr;
 
-	unsigned short int _countIndex = 0;
-	unsigned short* _aIndex = nullptr;
+	unsigned int _countIndex = 0;
+	unsigned int* _aIndex = nullptr;
 
 	bool _hasVBO = false;
 	unsigned int _buffer[4];
 
 public:
-	inline const unsigned short int countVertex()	{ return _countVertex; }
-	inline const unsigned short int countIndex()	{ return _countIndex; }
+	inline const unsigned int countVertex()	{ return _countVertex; }
+	inline const unsigned int countIndex()	{ return _countIndex; }
 
 	inline const float* const vertexes()			{ return _aVertex; }
 	inline const float* const normals()				{ return _aNormal; }
 	inline const float* const texCoords()			{ return _aTexCoord; }
-	inline const unsigned short* const indexes()	{ return _aIndex; }
+	inline const unsigned int* const indexes() { return _aIndex; }
 
 	inline bool hasVBO()						{ return _hasVBO; }
 	inline const unsigned int* const buffers()	{ return _buffer; }
@@ -40,9 +40,9 @@ public:
 	void setData(	float* const aVertex,
 					float* const aNormal,
 					float* const aTexCoord,
-					const unsigned short int countVertex,
-					unsigned short* const aIndex,
-					const unsigned short int countIndex);
+					const unsigned int countVertex,
+					unsigned int* const aIndex,
+					const unsigned int countIndex);
 
 public:
 	virtual ~Mesh();
